updates limit switch sensing for DriveToBar
This change updates the true/false nature of the limit switch sensing
for the DriveToBar command. Note that the limit switches are "false/low"
when they are open and "high/true" when they are closed.
This change also updates the speed of the robot when it is trying to
turn for the DriveToBar command after one switch has closed. The power
needed to be increased. Note that it may actually need to be amplified
for the turning to succeed. It takes more power to turn than it does to
drive forwards.