adjusts operator commands after testing
This commit adjusts various operator commands after testing them. The
intake and shooter motor spin directions were inverted so that a
positive value moves a ball through to the shooter and out of the robot.
A negative value moves a ball from the shooter to the conveyor and out
through the intake. The ExpelBalls, ShootBallsLow,
SpinConveyorBackwards, and SpinConveyorForwards commands were tested.