modifies auto code
This commit changes quite a few things. The major thing
it changes is it adds a timer to the end of DriveDistance
so that it does not turn of brakemode right after the distance
has been driven. It waits a certain amount of time to let the
brakemode brake the robot before ending the command. This also
completes the SimpleShootTwice auto command and allows it to
turn either left or right, depending on alliance needs. This also
stops spinning the intake while shooting. To compensate for this
the shooter speed is lowered since the motors have more current to
draw. Finally, there are a few vision changes.